chore: move traits used by legacy engine to separate mod
